agent: refator out Snapshot into a separate module
Add `evaluate()` for evaluation of a generic policy, while encapsulating
Snapshot refresh and retry semantics. We'll use this shortly for
additional authorization APIs.
