+++ crates/agent/src/api/authorize.rs
@@ -1,4 +1,4 @@
-use super::App;
+use super::{App, Snapshot};
 use anyhow::Context;
 use std::sync::Arc;
 
@@ -75,36 +75,19 @@ async fn do_authorize_task(app: &App, Request { token }: &Request) -> anyhow::Re
         }
     };
 
-    // Resolve the authorization snapshot against which this request is evaluated.
-    let snapshot = app.snapshot.read().unwrap();
-
-    let taken_unix = snapshot
-        .taken
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
-
-    // If the snapshot is too old then the client MUST retry.
-    if claims.iat > taken_unix + MAX_SNAPSHOT_INTERVAL.as_secs() {
-        begin_refresh(snapshot, &app.snapshot);
-
-        return Ok(Response {
-            retry_millis: jitter_millis(),
-            ..Default::default()
-        });
-    }
-
-    match evaluate_authorization(
-        &snapshot,
-        shard_id,
-        shard_data_plane_fqdn,
-        token,
-        &journal_name_or_prefix,
-        required_role,
-    ) {
+    match Snapshot::evaluate(&app.snapshot, claims.iat, |snapshot: &Snapshot| {
+        evaluate_authorization(
+            snapshot,
+            shard_id,
+            shard_data_plane_fqdn,
+            token,
+            &journal_name_or_prefix,
+            required_role,
+        )
+    }) {
         Ok((encoding_key, data_plane_fqdn, broker_address)) => {
             claims.iss = data_plane_fqdn;
-            claims.exp = claims.iat + exp_seconds();
+            claims.exp = claims.iat + super::exp_seconds();
 
             let token = jsonwebtoken::encode(&header, &claims, &encoding_key)
                 .context("failed to encode authorized JWT")?;
@@ -115,28 +98,11 @@ async fn do_authorize_task(app: &App, Request { token }: &Request) -> anyhow::Re
                 ..Default::default()
             })
         }
-        Err(err) if taken_unix > claims.iat => {
-            // The snapshot was taken AFTER the authorization request was minted,
-            // which means the request cannot have prior knowledge of upcoming
-            // state re-configurations, and this is a terminal error.
-            Err(err)
-        }
-        Err(_) => {
-            let retry_millis = if let Some(remaining) =
-                MIN_SNAPSHOT_INTERVAL.checked_sub(snapshot.taken.elapsed().unwrap_or_default())
-            {
-                // Our current snapshot isn't old enough.
-                remaining.as_millis() as u64
-            } else {
-                begin_refresh(snapshot, &app.snapshot);
-                0
-            } + jitter_millis();
-
-            Ok(Response {
-                retry_millis,
-                ..Default::default()
-            })
-        }
+        Err(Ok(retry_millis)) => Ok(Response {
+            retry_millis,
+            ..Default::default()
+        }),
+        Err(Err(err)) => Err(err),
     }
 }
 
@@ -281,205 +247,3 @@ fn evaluate_authorization(
         collection_data_plane.broker_address.clone(),
     ))
 }
-
-// Snapshot is a point-in-time view of control-plane state
-// that influences authorization decisions.
-pub struct Snapshot {
-    // Time immediately before the snapshot was taken.
-    taken: std::time::SystemTime,
-    // Platform collections, indexed on `journal_template_name`.
-    collections: Vec<SnapshotCollection>,
-    // Platform data-planes.
-    data_planes: tables::DataPlanes,
-    // Platform role grants.
-    role_grants: tables::RoleGrants,
-    // Platform tasks, indexed on `shard_template_id`.
-    tasks: Vec<SnapshotTask>,
-    // `refresh` is take()-en when the current snapshot should be refreshed.
-    refresh_tx: Option<futures::channel::oneshot::Sender<()>>,
-}
-
-// SnapshotCollection is the state of a live collection which influences authorization.
-// It's indexed on `journal_template_name`.
-struct SnapshotCollection {
-    journal_template_name: String,
-    collection_name: models::Collection,
-    data_plane_id: models::Id,
-}
-// SnapshotTask is the state of a live task which influences authorization.
-// It's indexed on `shard_template_id`.
-struct SnapshotTask {
-    shard_template_id: String,
-    task_name: models::Name,
-    spec_type: models::CatalogType,
-    data_plane_id: models::Id,
-}
-
-pub fn seed_snapshot() -> (Snapshot, futures::channel::oneshot::Receiver<()>) {
-    let (next_tx, next_rx) = futures::channel::oneshot::channel();
-
-    (
-        Snapshot {
-            taken: std::time::SystemTime::UNIX_EPOCH,
-            collections: Vec::new(),
-            data_planes: tables::DataPlanes::default(),
-            role_grants: tables::RoleGrants::default(),
-            tasks: Vec::new(),
-            refresh_tx: Some(next_tx),
-        },
-        next_rx,
-    )
-}
-
-pub async fn snapshot_loop(app: Arc<App>, mut refresh_rx: futures::channel::oneshot::Receiver<()>) {
-    while let Ok(()) = refresh_rx.await {
-        let (next_tx, next_rx) = futures::channel::oneshot::channel();
-        refresh_rx = next_rx;
-
-        match try_fetch_snapshot(&app.pg_pool).await {
-            Ok(mut snapshot) => {
-                snapshot.refresh_tx = Some(next_tx);
-                *app.snapshot.write().unwrap() = snapshot;
-            }
-            Err(err) => {
-                tracing::error!(?err, "failed to fetch snapshot (will retry)");
-                () = tokio::time::sleep(MIN_SNAPSHOT_INTERVAL).await;
-                _ = next_tx.send(()); // Wake ourselves to retry.
-            }
-        };
-    }
-}
-
-async fn try_fetch_snapshot(pg_pool: &sqlx::PgPool) -> anyhow::Result<Snapshot> {
-    tracing::info!("started to fetch authorization snapshot");
-    let taken = std::time::SystemTime::now();
-
-    let mut collections = sqlx::query_as!(
-        SnapshotCollection,
-        r#"
-            select
-                journal_template_name as "journal_template_name!",
-                catalog_name as "collection_name: models::Collection",
-                data_plane_id as "data_plane_id: models::Id"
-            from live_specs
-            where journal_template_name is not null
-            "#,
-    )
-    .fetch_all(pg_pool)
-    .await
-    .context("failed to fetch view of live collections")?;
-
-    let data_planes = sqlx::query_as!(
-        tables::DataPlane,
-        r#"
-            select
-                id as "control_id: models::Id",
-                data_plane_name,
-                data_plane_fqdn,
-                false as "is_default!: bool",
-                hmac_keys,
-                broker_address,
-                reactor_address,
-                ops_logs_name as "ops_logs_name: models::Collection",
-                ops_stats_name as "ops_stats_name: models::Collection"
-            from data_planes
-            "#,
-    )
-    .fetch_all(pg_pool)
-    .await
-    .context("failed to fetch data_planes")?;
-
-    let role_grants = sqlx::query_as!(
-        tables::RoleGrant,
-        r#"
-            select
-                subject_role as "subject_role: models::Prefix",
-                object_role as "object_role: models::Prefix",
-                capability as "capability: models::Capability"
-            from role_grants
-            "#,
-    )
-    .fetch_all(pg_pool)
-    .await
-    .context("failed to fetch role_grants")?;
-
-    let mut tasks = sqlx::query_as!(
-        SnapshotTask,
-        r#"
-            select
-                shard_template_id as "shard_template_id!",
-                catalog_name as "task_name: models::Name",
-                spec_type as "spec_type!: models::CatalogType",
-                data_plane_id as "data_plane_id: models::Id"
-            from live_specs
-            where shard_template_id is not null
-            "#,
-    )
-    .fetch_all(pg_pool)
-    .await
-    .context("failed to fetch view of live tasks")?;
-
-    let data_planes = tables::DataPlanes::from_iter(data_planes);
-    let role_grants = tables::RoleGrants::from_iter(role_grants);
-
-    // Shard ID and journal name templates are prefixes which are always
-    // extended with a slash-separated suffix. Avoid inadvertent matches
-    // over path component prefixes.
-    for task in &mut tasks {
-        task.shard_template_id.push('/');
-    }
-    for collection in &mut collections {
-        collection.journal_template_name.push('/');
-    }
-
-    tasks.sort_by(|t1, t2| t1.shard_template_id.cmp(&t2.shard_template_id));
-    collections.sort_by(|c1, c2| c1.journal_template_name.cmp(&c2.journal_template_name));
-
-    tracing::info!(
-        collections = collections.len(),
-        data_planes = data_planes.len(),
-        role_grants = role_grants.len(),
-        tasks = tasks.len(),
-        "fetched authorization snapshot",
-    );
-
-    Ok(Snapshot {
-        taken,
-        collections,
-        data_planes,
-        role_grants,
-        tasks,
-        refresh_tx: None,
-    })
-}
-
-fn begin_refresh<'m>(
-    guard: std::sync::RwLockReadGuard<'_, Snapshot>,
-    mu: &'m std::sync::RwLock<Snapshot>,
-) {
-    // We must release our read-lock before we can acquire a write lock.
-    std::mem::drop(guard);
-
-    if let Some(tx) = mu.write().unwrap().refresh_tx.take() {
-        () = tx.send(()).unwrap(); // Begin a refresh.
-    }
-}
-
-fn jitter_millis() -> u64 {
-    use rand::Rng;
-
-    // The returned jitter must always be positive.
-    // In production, it can take a few seconds to fetch a snapshot.
-    rand::thread_rng().gen_range(500..10_000)
-}
-
-fn exp_seconds() -> u64 {
-    use rand::Rng;
-
-    // Select a random expiration time in range [40, 80) minutes,
-    // which spreads out load from re-authorization requests over time.
-    rand::thread_rng().gen_range(40 * 60..80 * 60)
-}
-
-const MIN_SNAPSHOT_INTERVAL: std::time::Duration = std::time::Duration::from_secs(20);
-const MAX_SNAPSHOT_INTERVAL: std::time::Duration = std::time::Duration::from_secs(300); // 5 minutes.
